Fix unit tests for Ansible remediation generator
We need to update the compiled rule stored in the unit test suite data
(`tests/unit/ssg-module/data/file_owner_grub2_cfg.yml`) to follow the
current format of compiled rules. However, we want to test the
proper transformation of the "machine" platform in the rule definition
to the `when` statement. With the recent introduction of CPE platforms,
the class needs to be made aware of product's platforms which means
we now need to add the compiled platform and product to the test data.
Otherwise, the `when` statement would be empty.